fix HBaseTableAdmin parsing ttl from dataset spec

+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/dataset2/lib/table/hbase/HBaseTableAdmin.java
@@ -73,7 +73,7 @@ public class HBaseTableAdmin extends AbstractHBaseDataSetAdmin {
 
     String ttlProp = spec.getProperties().get(Table.PROPERTY_TTL);
     if (ttlProp != null) {
-      int ttl = Integer.parseInt(ttlProp);
+      long ttl = Long.parseLong(ttlProp);
       if (ttl > 0) {
         columnDescriptor.setValue(TxConstants.PROPERTY_TTL, String.valueOf(ttl));
       }
